move filter types to indexes module
