+++ ecoscope_workflows/indexes.py
@@ -1,0 +1,6 @@
+from typing import Literal, TypeAlias
+
+IndexName: TypeAlias = str
+IndexValue: TypeAlias = str
+Filter = tuple[IndexName, Literal["="], IndexValue]
+CompositeFilter = tuple[Filter, ...]
